% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.weights.R
\name{allocate.weights}
\alias{allocate.weights}
\title{Allocate Weights}
\usage{
allocate.weights(x, formaat = get("format", pos = 1), loc = "area",
  debug = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{Data frame containing survey data}

\item{formaat}{Character vector of the formats under consieration}

\item{loc}{Character vector referring to the variable containing different areas}

\item{debug}{Logical that assigns objects to the global environment for debugging purposes
when TRUE}

\item{verbose}{Logical that displays function messages when TRUE}
}
\description{
This function is used to allocate weights of various coal formats (such as a big or tin)
to different areas examined in the survey data
}
\details{
Make use of coalm.tables if you want the answer per suburb
}
\examples{
suburb.lst = lapply(levels(haq$town), FUN=function(x) {split(coalm,f=coalm[,x])[[2]] } )
names(suburb.lst) = levels(haq$town)
for (i in 1: length(suburb.lst)){
  suburb.lst[[i]][,"area"] <- names(suburb.lst)[[i]]
  }
res = lapply(suburb.lst, allocate.weights, formaat = haq$coal.buying.format)
}

