% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm.years.R
\name{bm.years}
\alias{bm.years}
\title{Basa Magogo Years}
\usage{
bm.years(haq, fix.text = FALSE, verbose = FALSE, fixdate = TRUE)
}
\arguments{
\item{haq}{Data frame containing survey data}

\item{fix.text}{Logical that replaces any text present for the year of initiation with the appropriate
integer and prints its summary information when TRUE}

\item{verbose}{Logical that displays function messages and summary information when TRUE}

\item{fixdate}{Logical that sets the reference date to the received date when TRUE}
}
\description{
This function is used to supply additional information in terms of variables such as the survey year,
the starting dates for BM users as well as the amount of winter and summer days that BM was used. It
can also create additional summary information relating to start and end dates as well as date
formatting capabilities. The HAQ survey data is edited and assigned to the global environment once
the function has completed.
}

