% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate.BM.manage.R
\name{calculate.BM.manage}
\alias{calculate.BM.manage}
\title{Calculate BM Management}
\usage{
calculate.BM.manage(x, verbose = FALSE, debug = FALSE, COEF = 2.3515,
  naampos = NULL)
}
\arguments{
\item{x}{a (readable binary-mode) connection or a
character string giving the name of the file to load}

\item{verbose}{displays function messages when TRUE}

\item{debug}{performs assignments to the global environment for
debugging purposes when TRUE}

\item{COEF}{the coefficient used for Emission calculation purposes}

\item{naampos}{the name position}
}
\value{
The output should be a melted data frame containing:

Number Households Actual

Percentage Coal Used by Household Actual

BASA Users Total end of vintage year Actual

Percentage Monitoring Loss Factor Actual

Percentage Nova BASA Users Retention Actual

Nova BASA Users Carried forward Actual

FACTOR VERS per Household existing Actual

FACTOR VERS per Household New Actual

Percentage New user start date Impact Actual

Low estimate Actual
}
\description{
Make a quick BM management summary (requires the fixname function)
}

