% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateER.AI.R
\name{calculateER.AI}
\alias{calculateER.AI}
\title{Calculate ER AI}
\usage{
calculateER.AI(haq = get("haq", envir = .GlobalEnv), pop = get("pop", envir
  = .GlobalEnv), fundir = "/Users/christiaanpauw/Dropbox/Rfunctions/",
  dropdir = "/Users/christiaanpauw/Dropbox/",
  thisplace = "Verifikasie2012/EMMnorth/",
  datadir = "/Users/christiaanpauw/Dropbox/Verifikasie2011/MalutiEast/data/",
  basedir = "/Users/christiaanpauw/Dropbox/Verifikasie2011/MalutiEast/",
  gendir = "/Users/christiaanpauw/Dropbox/Verifikasie2012/General_data/",
  dirr = "Rscripts/", tabdir = "tables/", name = "M11",
  projectname = "EMMnorth2011", write.date = "Sep11",
  reg.date = as.Date("2010-03-12"), verbose = 2, attend = att, ...)
}
\arguments{
\item{haq}{Data frame containing survey data (will be retreived from global environment)}

\item{pop}{Data frame containing town populations (will be retreived from global environment)}

\item{fundir}{The function directory as character vector}

\item{dropdir}{The dropbox directory as character vector}

\item{thisplace}{The directory of interenst for saving data as character vector}

\item{datadir}{The directory to save and retrieve data as character vector}

\item{basedir}{The directory containing base data as character vector}

\item{gendir}{The directory of the general data as character vector}

\item{dirr}{The directory containing R code as character vector}

\item{tabdir}{The directory to which tables should be saved as character vector}

\item{name}{The name to be attributed to the saved Rda file as character vector}

\item{projectname}{The project name that will be added to saved file names as character vector}

\item{write.date}{The date that files were written as character vector}

\item{reg.date}{The registration date as character vector}

\item{verbose}{Numerical that displays function messages when larger than 1}

\item{attend}{Data frame containing the attendance records}

\item{...}{Further arguments passed to or from other methods}
}
\description{
ER Calculation function for Alternatve Ignition Methodology
}
\references{
http://www.cdmgoldstandard.org/wp-content/uploads/2011/10/Alternative_Ignition.pdf
}

