% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalm.tables.R
\name{coalm.tables}
\alias{coalm.tables}
\title{Coal Merchant Survey (Main function posessing various internal functions)}
\usage{
coalm.tables(csv = NULL,
  fn = "~/Downloads/Coal_merchant_survey_20110620-100836.csv",
  manual.edit = NULL, manual.dest = "Bolata", selectvar = haq$town,
  fn.s = paste(basedir, tabdir, projectname, "coalm.Rda", sep = ""),
  tabdirr = paste(dropdir, basedir, tabdir, sep = ""), loc = haq$town,
  format = haq$coal.buying.format, debug = FALSE, verkort = TRUE,
  subplace = TRUE)
}
\arguments{
\item{csv}{If null the function chooses the csv option whereby coal merchant survey data is read
into R from the provided or default file name. If csv is not null the coal merchant
survey data is set to the provided data}

\item{fn}{The filename of coal merchant survey data to be used if read.csv is null}

\item{manual.edit}{Logical that initializes a manual edit when TRUE}

\item{manual.dest}{manual destination to be added to the town column}

\item{selectvar}{The selection variables to be used as a factor}

\item{fn.s}{The file name and directory to which the coal merchant data should be saved
as character vector}

\item{tabdirr}{The directory to which any tables should be saved as character vector}

\item{loc}{The locations under consideration as a factor}

\item{format}{The coal buying format information as a factor}

\item{debug}{Logical that assigns data to gobal environment for debugging process when TRUE}

\item{verkort}{Logical that shortens data by matching town data to selection variable data}

\item{subplace}{If TRUE the function melts and reconstructs the given data frame with
additional variables}

\item{x}{Numeric vecotr referring to the length of coal merchant data i.e the amount of merchants}

\item{data}{Data frame containing coal merchant information}

\item{format.idx}{The coal format under consideration}

\item{nr.measured}{The variable containing the numer of the format under consideration as character vector}

\item{x}{a vector (atomic or list) or an expression object.
Other objects (including classed objects) will be coerced by base::as.list}

\item{data}{Data frame containing coal merchant data}

\item{fromat.price}{The price of the format under consideration as character vector}

\item{mean}{weight The mean weight of the format under consideration as character vector}

\item{coalm}{Data frame containing coal merchant survey data}
}
\description{
Script to subset coal merchant survey for sepcific towns and create merchant an price latex tables.
}

