% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.bm.R
\name{code.bm}
\alias{code.bm}
\title{Basa Magogo coding}
\usage{
code.bm(haq = haq, coal.use = "coal.use", contact = "contact",
  verbose = FALSE)
}
\arguments{
\item{haq}{Data frame that contains the BM survey data}

\item{coal.use}{Character vector referring to the variable specifying whether a person
uses coal or not}

\item{contact}{Character vector referring to the variable containing contact information. Contains
either "yes" or "no"}

\item{verbose}{Logical taht displays function messages when TRUE}
}
\description{
This function is used to refine and manipulate any data contained in the data frame that relates to
BM usage. Once the function has been carried out the the edited data frame is assigned to the global environment
}

