% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.coal.R
\name{code.coal}
\alias{code.coal}
\title{Code Coal}
\usage{
code.coal(haq, verbose = FALSE, units = c("coal.units.winter.before.bm",
  "coal.units.summer.before.bm", "coal.units.winter.after.bm",
  "coal.units.summer.after.bm"), current.var = "current.coal.units.summer",
  summercoal.var = "coal.use.in.summer",
  summersavevar = "bm.saved.coal.summer",
  wintersavevar = "bm.saved.coal.winter", summersaveval = "No",
  wintersaveval = "No", estimate.summer = FALSE,
  summer.est.var = "bm.saved.coal.summer",
  summer.est.value = "It is not yet summer, or I only started recently and cannot tell exactly",
  summer.yes.value = "Yes", summer.no.value = "No",
  estimate.winter = FALSE, winter.est.var = "bm.saved.coal.winter",
  winter.est.value = "I have only started recently and cannot tell exactly")
}
\arguments{
\item{haq}{Data frame containing survey data}

\item{verbose}{Logical to display function messages}

\item{units}{Character vector referring to the variables that contain coal units used in
winter and summer before and after BM}

\item{current.var}{Character vector referring to the variable of current coal used in summer}

\item{summercoal.var}{The coal used in summer as character vector}

\item{summersavevar}{The variable containing data relating to coal and BM use during summer as
character vector}

\item{wintersavevar}{The variable containing data relating to coal and BM use during winter as
character vector}

\item{summersaveval}{The value relating to no savings during summer as character vector}

\item{wintersaveval}{The value relating to no savings during winter as character vector}

\item{estimate.summer}{Logical that makes use of summer estimation when TRUE}

\item{summer.est.var}{The summer estimation variable as character vector}

\item{summer.est.value}{The summer estimation value as character vector}

\item{summer.yes.value}{The value corresponding to yes for summer coal use as character vector}

\item{summer.no.value}{The value corresponding to no for summer coal use as character vector}

\item{estimate.winter}{Logical that makes use of winter estimation when TRUE}

\item{winter.est.var}{The winter estimation variable as character vector}

\item{winter.est.value}{The winter estimation value as character vector}
}
\description{
The main use of this function is to calculate the amount of savings to be attributed to
the BM method of lighting fires. The function modifies the provided data frame by adding
additional cost saving variables and assigning the completed data frame to the global environment
}

