% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.clm.row.R
\name{expand.clm.row}
\alias{expand.clm.row}
\title{Expand Coal Merchant Row}
\usage{
expand.clm.row(x = clm, levelvar = haq$town,
  del.var = "Delivery.suburbs_0", verbose = TRUE)
}
\arguments{
\item{x}{Data frame containing caol merchant survey information}

\item{levelvar}{The variable that is to have its levels accessed}

\item{del.var}{The delivery variable as character vector}

\item{verbose}{Logical that displays function messages when TRUE}
}
\description{
Function expands the coal merchant rows for every format - town combination
}
\details{
Use after clm$Delivery.suburbs_0 <- apply(clm,1, find.place.row)
}
\examples{
clm$town <- expand.clm.row(clm)
}

