% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCodebook.R
\name{extract.basevars}
\alias{extract.basevars}
\title{Extract Base Variables}
\usage{
extract.basevars(x, reflist, df.name = "var.name.df", lst.name = "lst",
  verbose = FALSE)
}
\arguments{
\item{x}{Data frame containing survey data}

\item{reflist}{A reference list}

\item{df.name}{Name of the data frame that is assigned to the global environment as character vector}

\item{lst.name}{Name of the list that is assigned to the global environment as character vector}

\item{verbose}{Displays function messages when TRUE}
}
\description{
Helper function to use for repeating variables e.g. person_1 , person_2 etc
}

