% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.place.row.R
\name{find.place.row}
\alias{find.place.row}
\title{Find Place Row}
\usage{
find.place.row(x = clm, del.string = "^Delivery.places_", cleardot = TRUE,
  spacer = ".", index = FALSE, levelvar = haq$town, verbose = TRUE,
  debug = FALSE)
}
\arguments{
\item{x}{The coal merchant survey data frame}

\item{del.string}{The string to be removed from certain variables as character vector}

\item{cleardot}{Logical that removes the punctuation marks from the list of places when TRUE}

\item{spacer}{a character vector where matches are sought, or an object which can be coerced by
as.character to a character vector}

\item{index}{Logical (not used here)}

\item{levelvar}{The level variable (not used here)}

\item{verbose}{Logical that displays function messages when TRUE}

\item{debug}{Logical to assign variables to the global environment for debugging purposes when TRUE}
}
\description{
Find the towns that are specified in the new coalm incorporation
}

