% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.table.v.R
\name{freq.table.by}
\alias{freq.table.by}
\title{Frequency table By}
\usage{
freq.table.by(x, by, tables.dir = tabdir)
}
\arguments{
\item{x}{The first variable to create the frequency table for}

\item{by}{The second variable used to group variable x by}

\item{tables.dir}{Directory to which any tables should be saved as character vector}
}
\description{
Creates a table from given information describing a value's frequency and percentage of the
total number of occurences for given variables and saves it to the provided directory
}
\details{
For the best table generation, assign the variable with more different values to
the "by" argument.
}

