% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.table.v.R
\name{freq.table.v}
\alias{freq.table.v}
\title{Frequency Table V}
\usage{
freq.table.v(x, tables.dir = tabdir)
}
\arguments{
\item{x}{Variable to create the frequency table for}

\item{tables.dir}{Directory to which all tables should be saved as character vector}
}
\description{
Creates a table from given information describing a value's frequency and percentage of the
total number of occurences for a given variable and saves it to the provided directory
}
\details{
The way this function is written, the name of the output .tex file will
contain the name of the label. If the variable has no label the output file will
not have a name.
}
\examples{
freq.table.v(haq$hh.BM.use, paste(thisplace,tabdir,sep=""))
}

