% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haq.labels.R
\name{haq.labels}
\alias{haq.labels}
\title{HAQ Labels}
\usage{
haq.labels(haq = haq)
}
\arguments{
\item{haq}{Data frame that contains survey data}
}
\description{
This function is used to create labels for all remaining variables contained in the HAQ survey data
frame considered to be in need of a label
}

