% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haq.labels.R
\name{initial.labels}
\alias{initial.labels}
\title{Initial HAQ Labeling}
\usage{
initial.labels(haq = haq)
}
\arguments{
\item{haq}{Data frame that contains household data}
}
\description{
This fuction creates labels for different variables contained in a data frame. It is used
early in the process for formats that are needed immediately.
}

