% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merchant.tables.R
\name{merchant.tables}
\alias{merchant.tables}
\title{Merchant Tables}
\usage{
merchant.tables(df = haq, coal.var = "coal.use", town.var = "town",
  merchant.var = "coal.merchant.r", datadir = datadir, limit = 5, n = 5,
  append = FALSE, xlsx = TRUE)
}
\arguments{
\item{df}{Data frame containing coal survey data}

\item{coal.var}{The variable referring to coal use as character vector}

\item{town.var}{The variable referring to towns as character vector}

\item{merchant.var}{The variable referring to coal merchants as character vector}

\item{datadir}{The path of the data directory as character vector}

\item{limit}{Numeric containing a limiting number to be used in output}

\item{n}{Numeric used in creating the top5 dataset}

\item{append}{If TRUE the data are appended to the connection when creating latex tables}

\item{xlsx}{If tRUE the fuction writes an xlsx document for the top5 dataset}
}
\description{
Function to make coal merchant tables by town
}

