% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.dates.R
\name{season.dates}
\alias{season.dates}
\title{Season Dates}
\usage{
season.dates(impyear = impyear)
}
\arguments{
\item{impyear}{The year of implementation as a numeric}
}
\description{
Function is supplied with a year of implementation, whereby it assigns values to the global environment
that indicate the start, middle and end as well as the number of days for winter and summer.
}

