% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitname.R
\name{splitname}
\alias{splitname}
\title{SplitName}
\usage{
splitname(x)
}
\arguments{
\item{x}{A character vector where matches are sought}
}
\description{
A useful function for splitting names with undesired formatting
}

