% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stitcher.R
\name{sticher}
\alias{sticher}
\title{Stitcher}
\usage{
sticher(indir = paste(dropdir, basedir, "1.ImplimentationData/rawdata/", sep =
  ""), outdir = paste(dropdir, basedir, "1.ImplimentationData/", sep = ""),
  projectname = "Seme2012", save = TRUE, xlsx = TRUE, verbose = FALSE,
  submissions = FALSE, write.csv = FALSE, clean = TRUE)
}
\arguments{
\item{indir}{The directory of the input files of concern for this function as character vector}

\item{outdir}{The directory to which any output files should be savesd as character vector}

\item{projectname}{The name of the project to be attributed to all of the desired output files}

\item{save}{Saves the HAQ data as an R data file in the output directory when TRUE}

\item{xlsx}{Creates and saves an xlsx document in the output directory when TRUE}

\item{verbose}{Displays function messages when TRUE}

\item{submissions}{creates a submissions index to go along with the selection index when TRUE}

\item{write.csv}{Writes a csv file of the HAQ data to the chosen output directory when TRUE}

\item{clean}{If TRUE acleaning function is initializes whereby certain variables are dropped from
the HAQ dataset}
}
\description{
Function is used to merge sheets
}

