% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.by.R
\name{table.by}
\alias{table.by}
\title{By Table Function}
\usage{
table.by(x)
}
\arguments{
\item{x}{The table to be manipulated}
}
\description{
Gives table without 0 counts. Very useful to use as function with by() or tapply()
}

