% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.AI.R
\name{test.AI}
\alias{test.AI}
\title{Alternative Ignition Data Intergity Test Function}
\usage{
test.AI(haq, pop)
}
\arguments{
\item{haq}{Data frame containing survey data}

\item{pop}{Data frame containing the populations for different locations}
}
\description{
This function is used to test data before running ER calculation (needs the fixname function)
}

