% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.coal.R
\name{unfactor}
\alias{unfactor}
\title{Unfactor}
\usage{
unfactor(x)
}
\arguments{
\item{x}{Data that will be converted}
}
\description{
Helper function for code.coal that creates a numeric vector from a factor
}

