% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merchant.tables.R
\name{yapply}
\alias{yapply}
\title{Helper Function Yapply}
\usage{
yapply(X, FUN, ...)
}
\arguments{
\item{X}{An atomic object}

\item{FUN}{function to apply}

\item{...}{optional arguments to FUN}
}
\description{
A function provided by Romain Francois based on something by Thomas Lumley, used in the creation of
merchant tables
}

